🐛 fix bug when to internet connection

+++ app/src/main/java/com/johnpaulcas/watchly/viewmodels/HomeViewModel.kt
@@ -1,5 +1,6 @@
 package com.johnpaulcas.watchly.viewmodels
 
+import android.util.Log
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -8,6 +9,7 @@ import androidx.lifecycle.viewModelScope
 import com.johnpaulcas.watchly.persistence.database.Track
 import com.johnpaulcas.watchly.persistence.database.TrackDao
 import com.johnpaulcas.watchly.repositories.home.HomeRepository
+import com.johnpaulcas.watchly.utils.NetworkUtil
 import com.johnpaulcas.watchly.utils.Resource
 import dagger.hilt.android.scopes.FragmentScoped
 import kotlinx.coroutines.launch
@@ -32,18 +34,26 @@ class HomeViewModel @ViewModelInject constructor(
 
     fun requestData() = viewModelScope.launch {
         _response.postValue(Resource.loading(null))
-        homeRepository.getTracks().let { response ->
-            if (response.isSuccessful) {
-                val tracks = response.body()?.results
-                tracks?.let {
-                    trackDao.deleteAll()
-                    trackDao.insertAll(it)
-                }
-                _response.postValue(Resource.success(null))
 
-            } else {
-                _response.postValue(Resource.error(response.errorBody().toString(), null))
+        val isInternetAvailable = NetworkUtil.isInternetAvailable()
+        // check internet availability
+        if (isInternetAvailable) {
+            homeRepository.getTracks().let { response ->
+                if (response.isSuccessful) {
+                    val tracks = response.body()?.results
+                    tracks?.let {
+                        trackDao.deleteAll()
+                        trackDao.insertAll(it)
+                    }
+                    _response.postValue(Resource.success(null))
+
+                } else {
+                    Log.d("LOGS", "requestData: ${response.errorBody().toString()}")
+                    _response.postValue(Resource.error(response.errorBody().toString(), null))
+                }
             }
+        } else {
+            _response.postValue(Resource.error("No internet :(", null))
         }
     }
